backend-common: Add runDockerContainer shared by Scaffolder and TechDocs
