+++ packages/backend-common/package.json
@@ -35,12 +35,14 @@
     "@backstage/integration": "^0.5.0",
     "@octokit/rest": "^18.0.12",
     "@types/cors": "^2.8.6",
+    "@types/dockerode": "^3.2.1",
     "@types/express": "^4.17.6",
     "archiver": "^5.0.2",
     "compression": "^1.7.4",
     "concat-stream": "^2.0.0",
     "cors": "^2.8.5",
     "cross-fetch": "^3.0.6",
+    "dockerode": "^3.2.1",
     "express": "^4.17.1",
     "express-promise-router": "^3.0.3",
     "fs-extra": "^9.0.1",
