+++ plugins/scaffolder-backend/src/scaffolder/stages/templater/helpers.ts
@@ -13,31 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Writable, PassThrough } from 'stream';
-import Docker from 'dockerode';
-import fs from 'fs';
-import { TemplateEntityV1alpha1 } from '@backstage/catalog-model';
 import { InputError } from '@backstage/backend-common';
+import { TemplateEntityV1alpha1 } from '@backstage/catalog-model';
 import { spawn } from 'child_process';
-
-export type RunDockerContainerOptions = {
-  imageName: string;
-  args: string[];
-  logStream?: Writable;
-  resultDir: string;
-  templateDir: string;
-  dockerClient: Docker;
-  createOptions?: Docker.ContainerCreateOptions;
-};
+import { PassThrough, Writable } from 'stream';
 
 export type RunCommandOptions = {
   command: string;
   args: string[];
   logStream?: Writable;
-};
-
-export type UserOptions = {
-  User?: string;
 };
 
 /**
@@ -88,81 +72,4 @@ export const runCommand = async ({
       return resolve();
     });
   });
-};
-
-/**
- *
- * @param options the options object
- * @param options.imageName the image to run
- * @param options.args the arguments to pass the container
- * @param options.logStream the log streamer to capture log messages
- * @param options.resultDir the /result path inside the container
- * @param options.templateDir the /template path inside the container
- * @param options.dockerClient the dockerClient to use
- */
-export const runDockerContainer = async ({
-  imageName,
-  args,
-  logStream = new PassThrough(),
-  resultDir,
-  templateDir,
-  dockerClient,
-  createOptions = {},
-}: RunDockerContainerOptions) => {
-  await new Promise<void>((resolve, reject) => {
-    dockerClient.pull(imageName, {}, (err, stream) => {
-      if (err) return reject(err);
-      stream.pipe(logStream, { end: false });
-      stream.on('end', () => resolve());
-      stream.on('error', (error: Error) => reject(error));
-      return undefined;
-    });
-  });
-
-  const userOptions: UserOptions = {};
-  // @ts-ignore
-  if (process.getuid && process.getgid) {
-    // Files that are created inside the Docker container will be owned by
-    // root on the host system on non Mac systems, because of reasons. Mainly the fact that
-    // volume sharing is done using NFS on Mac and actual mounts in Linux world.
-    // So we set the user in the container as the same user and group id as the host.
-    // On Windows we don't have process.getuid nor process.getgid
-    userOptions.User = `${process.getuid()}:${process.getgid()}`;
-  }
-
-  const [{ Error: error, StatusCode: statusCode }] = await dockerClient.run(
-    imageName,
-    args,
-    logStream,
-    {
-      Volumes: { '/result': {}, '/template': {} },
-      HostConfig: {
-        Binds: [
-          // Need to use realpath here as Docker mounting does not like
-          // symlinks for binding volumes
-          `${await fs.promises.realpath(resultDir)}:/result`,
-          `${await fs.promises.realpath(templateDir)}:/template`,
-        ],
-      },
-      ...userOptions,
-      // Set the home directory inside the container as something that applications can
-      // write to, otherwise they will just flop and fail trying to write to /
-      Env: ['HOME=/tmp'],
-      ...createOptions,
-    },
-  );
-
-  if (error) {
-    throw new Error(
-      `Docker failed to run with the following error message: ${error}`,
-    );
-  }
-
-  if (statusCode !== 0) {
-    throw new Error(
-      `Docker container returned a non-zero exit code (${statusCode})`,
-    );
-  }
-
-  return { error, statusCode };
 };
